Change type_arg to void * instead of unsigned int
For subdocuments, binary, or string types, type_arg should point to an unsigned
integer representing the size of the block of data. For array types, type_arg
should point to a 0-terminated array of integers representing the shape of the
NumPy array. For all other types, type_arg is ignored.